requirements.txt
accelerate==0.24.0
diffusers==0.21.4
google-cloud-storage==2.12.0
huggingface-hub==0.17.3
opencv-python==4.8.1.78
runpod==1.3.0
transformers==4.34.1